getting closer
There are exactly two messages per SRP session. The first is the
request (client-to-server). The second is the response (server-to-client).
The first must include H(KSession) to remind the server who we are. We
presume teh transport layer associates the response with the request, so we
don't need HK on the way back.